W5D1/Q3/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sentence_transformers==2.6.1
pinecone-client==3.2.2
openai==1.30.1
python-dotenv==1.0.0
pydantic==1.10.13
aiofiles==23.2.1
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
asyncpg==0.29.0 
pdfminer.six==20221105
beautifulsoup4==4.12.3 
prometheus-fastapi-instrumentator==6.0.0 